requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
bitarray==1.2.2
black==21.9b0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
colorclass==2.2.0
cytoolz==0.11.0
dataclasses==0.8
docopt==0.6.2
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.2
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.1
importlib-resources==5.4.0
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==5.2.0
mypy-extensions==0.4.3
netaddr==0.8.0
packaging==21.2
parsimonious==0.8.1
pathspec==0.9.0
pip-upgrader==1.4.15
pipreqs==0.4.11
pkg-resources==0.0.0
platformdirs==2.4.0
protobuf==3.19.1
py-solc==3.2.0
py-solc-x==1.1.1
pycryptodome==3.11.0
pyparsing==2.4.7
pyrsistent==0.18.0
regex==2021.10.23
requests==2.26.0
rlp==2.0.1
semantic-version==2.8.5
six==1.16.0
terminaltables==3.1.0
tomli==1.2.2
toolz==0.11.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
upgrade-requirements==1.7.0
urllib3==1.26.7
varint==1.0.2
web3==5.24.0
websockets==9.1
yarg==0.1.9
yarl==1.7.0
zipp==3.6.0
